Feat: 코드 리스트부분 뷰와 뷰모델 분리 + 네비게이션바 추가
